requirements.txt
accelerate
aiohttp
aiosignal
altair
astor
asttokens
async-timeout
attrs
autopep8
backcall
blinker
cachetools
certifi
charset-normalizer
click
colorama
contourpy
cycler
decorator
entrypoints
et-xmlfile
executing
filelock
Flask
fonttools
frozenlist
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-auth
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
huggingface-hub
idna
importlib-metadata
ipython
itsdangerous
jedi
Jinja2
jsonschema
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
multidict
numpy
openai
openpyxl
packaging
pandas
pandas-stubs
pandasai==0.2.14
parso
pickleshare
Pillow
pip
prompt-toolkit
proto-plus
protobuf
psutil
pure-eval
pyarrow
pyasn1
pyasn1-modules
pycodestyle
pydeck
Pygments
Pympler
pyparsing
pyrsistent
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
rich
rsa
setuptools
six
smmap
stack-data
streamlit
tenacity
tokenizers
toml
toolz
torch
tornado
tqdm
traitlets
transformers
typing_extensions
tzdata
tzlocal
urllib3
validators
watchdog
wcwidth
Werkzeug
yarl
zipp
